+++ generic-table/interfaces/gt-config-setting.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=gt-config-setting.js.map