+++ generic-table/interfaces/gt-options.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=gt-options.js.map