+++ generic-table/interfaces/gt-row.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=gt-row.js.map