+++ generic-table/interfaces/gt-texts.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=gt-texts.js.map