+++ index.js
@@ -1,0 +1,10 @@
+"use strict";
+function __export(m) {
+    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
+__export(require('./generic-table/components/gt-expanding-row.component'));
+__export(require('./generic-table/components/generic-table.component'));
+__export(require('./generic-table/components/gt-table-info.component'));
+__export(require('./generic-table/components/gt-pagination.component'));
+__export(require('./generic-table/generic-table.module'));
+//# sourceMappingURL=index.js.map