+++ src/generic-table/components/gt-table-info.component.ts
@@ -4,7 +4,7 @@ import {GenericTableComponent} from './generic-table.component';
 
 @Component({
   selector: 'gt-table-info',
-  template: `<span>{{(customText? customText:genericTable.gtTexts) | gtTableInfo:genericTable.gtInfo:genericTable.gtInfo.recordsAfterSearch:genericTable.gtInfo.recordFrom:genericTable.gtInfo.recordTo}}</span>`
+  template: `<span *ngIf="genericTable.gtInfo" >{{(customText? customText:genericTable.gtTexts) | gtTableInfo:genericTable.gtInfo:genericTable.gtInfo.recordsAfterSearch:genericTable.gtInfo.recordFrom:genericTable.gtInfo.recordTo:genericTable.gtInfo.recordsAll}}</span>`
 })
 export class GtTableInfoComponent implements AfterViewChecked {
 
@@ -23,18 +23,17 @@ import { Pipe, PipeTransform } from '@angular/core';
 import { GtInformation } from '../interfaces/gt-information';
 import { GtTexts } from '../interfaces/gt-texts';
 
-
 @Pipe({
   name: 'gtTableInfo'
 })
 export class TableInfoPipe implements PipeTransform {
 
   transform(texts:GtTexts,keys:GtInformation,refresh:number): string {
-    let text:string = texts.tableInfo;
+    let text = texts.tableInfo;
     if(keys.recordsAfterSearch !== keys.recordsAll){
       text = texts.tableInfoAfterSearch;
     }
-    for (var key in keys) {
+    for (let key in keys) {
       if( keys.hasOwnProperty(key) ) {
         let searchString = new RegExp('#'+key, 'g');
         text = text.replace(searchString,keys[key]);